chore: split combobox and searchbox
